corrigindo os relacionamentos many to many e terminando de implementar autorização de usuários

+++ app/Http/Controllers/ProjectController.php
@@ -24,12 +24,15 @@ class ProjectController extends Controller
 
     public function index()
     {
-        return $this->repository->with(["users", "clientes"])->all();
+        return $this->repository->with(["users", "clientes"])->findWhere(array("owner_id" => Authorizer::getResourceOwnerId()));
     }
     
     public function members($id)
     {
-        return $this->members_repository->with(["users"])->findByField('project_id', $id);
+        $project = $this->repository->find($id);
+        return $project->members;
+        //return $this->members_repository->with(["users"])->findByField('project_id', $id);
+        //qual é a melhor forma?
     }    
 
     public function store(Request $request)
@@ -38,10 +41,8 @@ class ProjectController extends Controller
     }
 
     public function show($id)
-    {     
-        $user_id = Authorizer::getResourceOwnerId();
-        
-        if($this->repository->isOwner($id, $user_id) == false)
+    {             
+        if($this->checkProjectPermissions($id) == false)
         {
             return ["success" => false];
         }        
@@ -51,11 +52,21 @@ class ProjectController extends Controller
 
     public function update(Request $request, $id)
     {
+        if($this->isOwner($id) == false)
+        {
+            return ["success" => false];
+        }          
+        
         return $this->service->update($request->all(), $id);
     }
 
     public function destroy($id)
     {
+        if($this->isOwner($id) == false)
+        {
+            return ["success" => false];
+        }          
+        
         $this->repository->delete($id);
     }
     
@@ -71,9 +82,32 @@ class ProjectController extends Controller
               
     }
     
+    //rever este método, tendo em vista o hasMember abaixo
     function isMember($id, $userId) {
         
         return $this->service->isMember($id, $userId); 
               
-    }   
+    }
+    
+    function isOwner($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();        
+        return $this->repository->isOwner($projectId, $userId);        
+    }
+    
+    function hasMember($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();        
+        return $this->repository->hasMember($projectId, $userId);        
+    } 
+    
+    function checkProjectPermissions($projectId)
+    {
+        if($this->isOwner($projectId) || $this->hasMember($projectId))
+        {
+            return true;
+        }
+        
+        return false;
+    }
 }
